requirements.txt
Django~=4.2.20
google-cloud-pubsub>=2.0.0
google-cloud-logging>=3.0.0
psycopg2
firestore